requirements.txt.txt
streamlit==1.37.0
pandas==2.2.2
numpy==1.26.4
plotly==5.23.0
python-dotenv==1.0.1
openpyxl==3.1.5
python-pptx==0.6.23
kaleido==0.2.1
statsmodels==0.14.2
jinja2==3.1.4
# IA (comentarios automáticos)
google-genai==0.3.0
# PDF opcional: en Streamlit Cloud puede requerir más dependencias; si te falla, comenta las 2 de abajo
weasyprint==62.3
reportlab==4.2.2
